Made a login container component and moved the form to presentational
